requirements.txt
anthropic>=0.92.0
PyYAML>=6.0
